+++ lib/src/details/components/body.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:oldwave/src/models/product.dart';
+
+class Body extends StatelessWidget {
+  final Product product;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      
+    );
+  }
+}
